fix(profile): Load profile
- No wallet connected view
