+++ src/components/Menu/index.tsx
@@ -4,7 +4,7 @@ import { useWallet } from '@binance-chain/bsc-use-wallet'
 import { allLanguages } from 'config/localisation/languageCodes'
 import { LanguageContext } from 'contexts/Localisation/languageContext'
 import useTheme from 'hooks/useTheme'
-import { usePriceCakeBusd } from 'state/hooks'
+import { usePriceCakeBusd, useProfile } from 'state/hooks'
 import config from './config'
 
 const Menu = (props) => {
@@ -12,6 +12,7 @@ const Menu = (props) => {
   const { selectedLanguage, setSelectedLanguage } = useContext(LanguageContext)
   const { isDark, toggleTheme } = useTheme()
   const cakePriceUsd = usePriceCakeBusd()
+  const { profile } = useProfile()
 
   return (
     <UikitMenu
@@ -25,6 +26,12 @@ const Menu = (props) => {
       setLang={setSelectedLanguage}
       cakePriceUsd={cakePriceUsd.toNumber()}
       links={config}
+      profile={{
+        username: profile?.username,
+        image: `/images/nfts/${profile?.nft.previewImage}`,
+        profileLink: '/profile',
+        noProfileLink: '/profile',
+      }}
       {...props}
     />
   )
