+++ src/views/Profile/components/ComingSoon.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Flex, Heading } from '@pancakeswap-libs/uikit'
+import { BunnyPlaceholderIcon, Flex, Heading } from '@pancakeswap-libs/uikit'
 import useI18n from 'hooks/useI18n'
 
 interface ComingSoonProps {
@@ -11,7 +11,7 @@ const ComingSoon: React.FC<ComingSoonProps> = ({ children }) => {
 
   return (
     <Flex flexDirection="column" alignItems="center" justifyContent="center" p="24px">
-      <img src="/images/bunny-placeholder.svg" alt="Bunny Placeholder" />
+      <BunnyPlaceholderIcon width="72px" height="72px" />
       <Heading as="h5" size="md" color="textDisabled">
         {children || TranslateString(999, 'Coming Soon!')}
       </Heading>
