+++ src/views/Profile/components/Header.tsx
@@ -1,7 +1,6 @@
-import { Button, Heading } from '@pancakeswap-libs/uikit'
+import { Heading } from '@pancakeswap-libs/uikit'
 import useI18n from 'hooks/useI18n'
 import React from 'react'
-import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 
 const StyledProfileHeader = styled.div`
